// src/api/generated/model/driverResponseStatus.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * MiniCRM API
 * API Documentation for MiniCRM Application
 * OpenAPI spec version: 1.0.0
 */

/**
 * Current employment status
 */
export type DriverResponseStatus = typeof DriverResponseStatus[keyof typeof DriverResponseStatus];


export const DriverResponseStatus = {
  ACTIVE: 'ACTIVE',
  DELETED: 'DELETED',
} as const;